fix lora communication
- use unspcifed protocol instead of tcp which was wrong
- send data as byte array of floats and bools instead of json because
  maximum allowed is 58 bytes

+++ EV Car UI/Models/CanBus.cs
@@ -88,8 +88,8 @@ public class RandomData : IDataReceiver
                     GetRandomFloat(25),
                     GetRandomFloat(60),
                     GetRandomFloat(5000),
-                    Increment * (loopsCompleted - loopsToWaitBeforeStarting) % 100,
-                    Increment * (loopsCompleted - loopsToWaitBeforeStarting) % 100,
+                    (float) Increment * (loopsCompleted - loopsToWaitBeforeStarting) % 100,
+                    (float) Increment * (loopsCompleted - loopsToWaitBeforeStarting) % 100,
                     GetRandomBool(),
                     GetRandomBool(),
                     GetRandomBool()));
